Add prettier format task, refactor files
