+++ packages/create-typink/src/tasks/index.ts
@@ -2,3 +2,5 @@ export * from './createFirstCommit.js';
 export * from './copyTemplateFiles.js';
 export * from './createProjectDirectory.js';
 export * from './installPackages.js';
+export * from './prettierFormat.js';
+export * from './createProject.js';
