+++ packages/create-typink/src/utils/index.ts
@@ -1,5 +1,3 @@
-export * from './createProject.js';
-export * from './promptOptions.js';
-export * from './parseArguments.js';
 export * from './render.js';
 export * from './string.js';
+export * from './options.js';
